Shield Java NIO layer from usage patterns that can deteriorate the overall latency.
It seems that copying data to the intermediate buffer is cheaper than invoking channel write
too frequently. This is a conclusion after profiling using RoundTripTimeTest::shouldExchangeMessagesAtHigherRate

When running after the change was applied, it is still possible to achieve reasonable latency at a rate of 400 000 messages per second (500 000 messages per second requires further optimization or a faster box than mine to keep the rond trip latencies low)

Sending 400 000 msg/s results:
---------------------------------------------------------
latency (microseconds) |     ~ one way |     round trip |
mean                   |            12 |             23 |
99th percentile        |            26 |             51 |
99.9th percentile      |            55 |            109 |
99.99th percentile     |            92 |            184 |
99.999th percentile    |           100 |            199 |
worst                  |           100 |            199 |

To reproduce:

java -classpath "./asynctcpacceptance/build/distributions/asynctcpacceptance/lib/*" dev.squaremile.asynctcpacceptance.EchoApplication 9998

java -classpath "./asynctcpacceptance/build/distributions/asynctcpacceptance/lib/*" dev.squaremile.asynctcpacceptance.SourcingConnectionApplication localhost 9998 400000 10000 5000000 100 1 0

+++ asynctcpacceptance/src/test/java/dev/squaremile/asynctcpacceptance/ServerReceivesConnectionsTest.java
@@ -39,7 +39,7 @@ class ServerReceivesConnectionsTest extends TransportTestBase
     void shouldNotifyWhenConnected()
     {
         // Given
-        serverTransport.handle(serverTransport.command(Listen.class).set((long)1, freePort(), rawStreaming()));
+        serverTransport.handle(serverTransport.command(Listen.class).set(1, freePort(), rawStreaming()));
         final int serverPort = serverTransport.events().last(StartedListening.class).port();
         assertThat(serverTransport.statusEvents().contains(NumberOfConnectionsChanged.class)).isFalse();
 
@@ -70,7 +70,7 @@ class ServerReceivesConnectionsTest extends TransportTestBase
     void shouldProvideConnectionDetailsForEachConnection()
     {
         // Given
-        serverTransport.handle(serverTransport.command(Listen.class).set((long)5, freePort(), rawStreaming()));
+        serverTransport.handle(serverTransport.command(Listen.class).set(5, freePort(), rawStreaming()));
         final int serverPort = serverTransport.events().last(StartedListening.class).port();
 
         // When
@@ -92,7 +92,7 @@ class ServerReceivesConnectionsTest extends TransportTestBase
     void shouldCloseConnection()
     {
         // Given
-        serverTransport.handle(serverTransport.command(Listen.class).set((long)9, freePort(), rawStreaming()));
+        serverTransport.handle(serverTransport.command(Listen.class).set(9, freePort(), rawStreaming()));
         serverTransport.workUntil(() -> !serverTransport.events().all(StartedListening.class).isEmpty());
         final int serverPort = serverTransport.events().last(StartedListening.class).port();
         assertThrows(SocketException.class, clients.client(1)::write); // throws if not connected when writing
@@ -159,8 +159,8 @@ class ServerReceivesConnectionsTest extends TransportTestBase
         // Given
         final int listeningPort1 = freePort();
         final int listeningPort2 = freePortOtherThan(listeningPort1);
-        serverTransport.handle(serverTransport.command(Listen.class).set((long)5, listeningPort1, rawStreaming()));
-        serverTransport.handle(serverTransport.command(Listen.class).set((long)6, listeningPort2, rawStreaming()));
+        serverTransport.handle(serverTransport.command(Listen.class).set(5, listeningPort1, rawStreaming()));
+        serverTransport.handle(serverTransport.command(Listen.class).set(6, listeningPort2, rawStreaming()));
         assertThat(serverTransport.events().last(StartedListening.class, event -> event.commandId() == 5).port()).isEqualTo(listeningPort1);
         assertThat(serverTransport.events().last(StartedListening.class, event -> event.commandId() == 6).port()).isEqualTo(listeningPort2);
 
@@ -195,10 +195,10 @@ class ServerReceivesConnectionsTest extends TransportTestBase
         final int clientPort2 = freePortOtherThan(listeningPort1, listeningPort2, clientPort1);
 
         // When
-        serverTransport.handle(serverTransport.command(Listen.class).set((long)5, listeningPort1, rawStreaming()));
+        serverTransport.handle(serverTransport.command(Listen.class).set(5, listeningPort1, rawStreaming()));
         assertThat(serverTransport.events().last(StartedListening.class, event -> event.commandId() == 5).port()).isEqualTo(listeningPort1);
         serverTransport.workUntil(completed(() -> clients.client(1).connectedTo(listeningPort1, clientPort1)));
-        serverTransport.handle(serverTransport.command(Listen.class).set((long)6, listeningPort2, rawStreaming()));
+        serverTransport.handle(serverTransport.command(Listen.class).set(6, listeningPort2, rawStreaming()));
         assertThat(serverTransport.events().last(StartedListening.class, event -> event.commandId() == 6).port()).isEqualTo(listeningPort2);
         serverTransport.workUntil(completed(() -> clients.client(2).connectedTo(listeningPort2, clientPort2)));
